Improve merging for certain entry types.
The resource merging should not just be combining the files in each
dependency file set, but rather should be using the actual AAPT resource
merging APIs to accomplish this.

This also improves the merging process for JARs as well, allows merging
identical files but also allows for declaring conflicts or skips of
files with the same name but different contents.