Allow merging multiple entries at once.
For certain entries in the final merged AAR, such as R.txt or resource
files, there are fast, efficient ways to merge all of the dependency
archives' resource files at the same time, instead of a series of
synchronous folds of two at a time. The result is that certain merges
can be done in O(1) instead of O(N), making the merging step of the
packaging task much faster.