Fix dependency matching drastically.
One critical issue that was not being addressed by the sample code this
was running on is that additional logic was required for depending on
other projects. Selection and disambiguation of the proper artifact type
was necessary to retrieve the appropriate AAR or JAR file from consumed
library module dependencies.

This also turns the transitivity of the publish<Flavor>Classpath off,
which greatly (and properly!) limits the scope of dependencies being
processed to only that of what's being specified in the appropriate
publish configurations.