Merge resources in reverse.
Adding DataSets in the reverse order allows the later ones to take
precedence over the earlier-added ones, which is how Android resource
merging normally works. This allows dependencies "higher" in the
classpath to have priority over those "lower" in the classpath when it
comes to resource declarations, allowing the final library/application
to have the final say on resource values.