Capture missing API element modules.
Dependencies declared as api(...) dependencies of the module being
packaged would be declares as runtime dependencies, *not* compile,
in the generated POM file, because they were not identified as being
exported as an outgoing compilation dependency for this module.

We were only checking the variant-specific `*ApiDependenciesMetadata`
configuration for api dependencies, which I incorrectly assumed would
contain all of the variant-agnostic dependencies as well.