Introduce AGP Compatibility layer for plugin.
We cannot predict exactly which version of AGP will be on our classpath
at runtime, and we want to be able to support at least a few different
major versions at a time, some of which may be binary-incompatible with
one another and have breaking API changes. Because of this, we need to
create a facade in front of any AGP-defined type, using our own classes
in order to create the same kind of behaviour across different versions.

Closes issue #4.