Transform AAR using artifact transform pipeline.
Newer versions of AGP allow certain transformations of artifacts to be
registered per variant:
- Adding/appending new files
- Mutating files
- Replacing files or moving them around

This is done per `com.android.build.api.artifact.Artifact`, and we can
wire up the PackageAarTask to transform the outputted AAR *during* the
assemble process, instead of after the fact. This allows us to cut out a
lot of the custom work creating new configurations and components in
order to allow maven-publish to find the files and metadata.

Developers can choose which variants enable aaraar packaging using the
existing `aaraar` extension:

```
aaraar {
  isEnabledForVariant { variant ->
    variant.name == "publish"
  }
}
```