Move merging logic into isolated classes.
The data models for individual entries in a JAR/AAR file no longer
define how they are meant to be merged themselves. Instead, this
responsibility now belongs to the Merger<*> interface, which defines how
multiple instances of the same entry are merged with one another.
This should allow for better testability and customization or additional
validation of how entries are merged together.