Compatibility with Java modules.
`aaraar` can now be applied to non-Android modules, particularly any
module that applies the `java` plugin, either directly or transitively,
is now capable of producing a merged `jar` file. It works by depending
on the jar task in the project, and creating the merged jar file
in-place to replace the pre-existing output.

Additional features like discarding META-INF files and class shading are
supported out of the box.