Migrate to a post-processor factory pattern.
The ApiJarProcessor ran after the merged archive was prepared, but that
does not mean that consumers may wish to implement other transformations
or inspections on the merged archive. Making this process more generic,
while still keeping an interface for making adding an `api.jar` to the
merged archive, achieves this.