Set the `org.gradle.jvm.environment` capability.
In cases where a module produces a JVM and Android artifact, we want
consuming Android projects to match against the Android variant, not the
JVM variant. Adding an AttributeCompatibilityRule for the variant's
TargetJvmEnvironment is not possible since there is already one
registered by the ecosystem plugin that marks all values as compatible.